A lot of improvements about crossbrowsing in the player
We now ask the browser if it supports webm(vp8/vorbis), ogg(theora/vorbis) or
mp4(avc1/mp4). If none of that formatas is supported, a flash player will be
launched - if we have a flv source.

If we find IE9, that _says_ that support mp4, we force the flash player.